Extract all array helper functionality into its own internal module
